rename tokenUsdAmount & change parameter and state
